+++ qiperfc/src/tpplot.h
@@ -17,7 +17,7 @@ class TPPlot : public QCustomPlot
 public:
     explicit TPPlot(QWidget *parent = nullptr);
     void setStartTime(QDateTime startTime);
-    void addTPData(QString idx, double xdata, double ydata, int lostrate);  //
+    void addTPData(QString idx, double xdata, double ydata, double lostrate);  //
     QCPGraph *getGraph(QString idx); // get QCPGraph by index
     QCPBars *getLostRateGraph(QString idx);
     void clear();
